Cleanup examples and add static components.
Static components aren't finished, but they are ready to fill in.
These are the simplest part of the web app so they should be taken
care of first.

After the static components are finished I'll go through and remove
anything that is unnecessary (it will be in the git history anyways).

Then I'll be ready to work on the dynamic content.