Data detail page pulls after route is fetched.
Okay... this stuff is hacky but it works for now.

Previously, I was dispatching actions to pull the proteins from the
API prior to requesting the link for the protein detail. This worked
fine when navigating from /data, but made it impossible to go to the
page by typing venomkb.org/{venomkb_id} directly into the url bar.

I moved the actions into ProteinDetailContainer.componentWillMount(),
but then it was trying to render parts of the record prior to actually
receiving the data from the fetch() call. I 'fixed' this by adding
extra conditionals in render()...

So, something is still wrong with 'isFetching', but I don't know
exactly what. In order to truly fix this once and for all, I'll have
to change isFetching to accurately report whether the requested datum
has been received or not.