Common actions for selecting and fetching species and proteins
Instead of maintaining separate actions for proteins and species, I rolled them into selectData, fetchData, etc.

There is a conditional in fetchData that constructs the correct API route for the requested data type, but other than that there are no real wrinkles in the new implementation. I /think/ that renaming ‘proteins’ to ‘data’ is finished, but some other inconsistencies may pop up.

Next: fill in the details for rendering ‘Species’