Add species images.
Includes python scripts to find the images on wikimedia.

DON'T RUN THESE AGAIN UNLESS NECESSARY! If you absolutely have to,
here's how to do it: (1) Run `fetch_images_wikidata.py` to search for
images on wikidata, and download them to index/img/species. Also
writes species_image_map.pkl to the scripts directory. (2) Go through
the species images, and delete any that are bad (i.e., that they refer
to something else). (3) Run `deploy_species_image_urls.py` to write
the urls to the non-deleted images to the database. I know... this is
a weird way to do it, but made it easy to review the images locally
while not storing them (yet) on the server.