Render more small details; prepare distribution webpack.
This is a hodge-podge of little things. First, as the title line
indicates, there are some new little additions to the rendering, such
as a dropdown for running BLASTp on protein sequences.

venomkb.html now includes a link to minified jquery, which is required
for the pulldown on the About button. Speaking about the "About"
button, I need to finish rendering the components it contains.

The big (and yet unfinished) task is adding the express and webpack
configuration to prep the site for deployment. Most of what I did is
from the following site:

https://medium.com/@patriciolpezjuri/using-create-react-app-with-react-router-express-js-8fa658bf892d

Like I said, it's still unfinished. I've figured out how to build the
images to index/dist, but the site can't find the entry point for the
react app itself yet.