Rewrite DB index api and redux code
Big changes for one commit... Instead of querying for important fields
within the proteins and species collections, I've created a separate
collection called Dbindexitems that stores all of this in a static
format. The collection can be regenerated any time it's necessary by
running index/scripts/build_mongo_index.py.

Also, genomes are now passed to DataVirtualized, but they aren't
yet showing up in the list (need to add filtering conditions in order for
that to work).

Additionally, there is now decent support for literature references,
which aren't yet rendering but are in the redux store.