Work towards minifying production code.
Not quite there, as the title implies. I'll come back to it later -
basically I need to turn off the devtool directive in webpack so it
doesn't write source maps into main.js whenever i build the site, or
only have it do so when in development mode. However, it seemed to
break VenomKB somehow when I tried this. This is probably going to be
the first thing i come back to after completing the MVP.