Reconcile buggy production with local changes.
Okay this was a doozie... Everything worked when developing locally,
even when i was in production mode, but when I tried to build on the
production server the React component refused to render. After a ton of
debugging (refer to 'debug-production' branch), I was able to restore
functionality by reverting code in index/containers/Root.js. This leads
me to believe that there is something somewhere that relies on my
erroneously defined "process.env.NODE_ENVIRONMENT" variable, instead of
the expected process.env.NODE_ENV. This will be an issue to explore
after the site is working (along with the dev source map issue and
whatnot).